GrantIQ-WebScrapper/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
pymongo==4.6.2
python-dotenv==1.0.1
Flask==3.0.2
google-generativeai==0.4.1
selenium
undetected-chromedriver
playwright
playwright-stealth
setuptools